+++ src/ci.hpp
@@ -92,7 +92,7 @@ namespace ci {
 			const size_t mmax,
 			const std::string path = std::string(getenv("HOME")) + "/.crypto-incognito/mG.bin"):
 			mmax(mmax), mG(mmax) {
-			int elemsRead = ci_ecelgamal_load_mg(this->mG.data(), mmax, path.c_str());
+			size_t elemsRead = ci_ecelgamal_load_mg(this->mG.data(), mmax, path.c_str());
 			if(elemsRead != mmax) throw "Failed to load mG.bin.";
 		}
